* 'master' of github.com:phretor/oss-fuzz: (144 commits)
  [infra] Add different messages depending on bisection type (#6547)
  Fix false positive result due to floating point divide by zero (which is not a bug). (#6549)
  rocksdb: fix build (#6546)
  dnsmasq: fix fuzzer and build (#6545)
  sudoers: set optimization flag in LDFLAGS for the object size sanitizer (#6544)
  [clusterfuzzlite] Keep unaffected fuzzers when uploading a build. (#6530)
  [clusterfuzzlite] Include shebang line when generating project. (#6532)
  suricata: default rules in out directory (#6540)
  protobuf-java: update contact information (#6542)
  update tendermint/tendermint project file (#6541)
  [infra] Add block listed target name logic from ClusterFuzz (#6538)
  Done (#6536)
  build(deps): bump nokogiri from 1.12.2 to 1.12.5 in /docs (#6535)
  Add ddiproietto@google.com to CC (#6533)
  Fix for rust and swift coverages (#6517)
  [bls-signatures] Enable additional operations (#6522)
  binutils: fix build (#6520)
  Undefined name: import sys for line 33 (#6519)
  Update project.yaml
  mupdf: fix build (#6518)
  ...

+++ projects/binutils/build.sh
@@ -60,13 +60,16 @@ for i in readelf; do
 done
 
 # Link the files
-## Readelf
-$CXX $CXXFLAGS $LIB_FUZZING_ENGINE -W -Wall -I./../zlib -o fuzz_readelf fuzz_readelf.o version.o unwind-ia64.o dwarf.o elfcomm.o ../libctf/.libs/libctf-nobfd.a -L/src/binutils-gdb/zlib -lz ../libiberty/libiberty.a 
-mv fuzz_readelf $OUT/fuzz_readelf
+# Only link if they exist
+if ([ -f dwarf.o ] && [ -f elfcomm.o ] && [ -f version.o ]); then
+  ## Readelf
+  $CXX $CXXFLAGS $LIB_FUZZING_ENGINE -W -Wall -I./../zlib -o fuzz_readelf fuzz_readelf.o version.o unwind-ia64.o dwarf.o elfcomm.o ../libctf/.libs/libctf-nobfd.a -L/src/binutils-gdb/zlib -lz ../libiberty/libiberty.a 
+  mv fuzz_readelf $OUT/fuzz_readelf
 
-### Set up seed corpus for readelf in the form of a single ELF file. 
-zip fuzz_readelf_seed_corpus.zip /src/fuzz_readelf_seed_corpus/simple_elf
-mv fuzz_readelf_seed_corpus.zip $OUT/ 
+  ### Set up seed corpus for readelf in the form of a single ELF file. 
+  zip fuzz_readelf_seed_corpus.zip /src/fuzz_readelf_seed_corpus/simple_elf
+  mv fuzz_readelf_seed_corpus.zip $OUT/ 
 
-## Copy over the options file
-cp $SRC/fuzz_readelf.options $OUT/fuzz_readelf.options
+  ## Copy over the options file
+  cp $SRC/fuzz_readelf.options $OUT/fuzz_readelf.options
+fi
